requirements.txt
fastapi
uvicorn[standard]                
sqlalchemy>=1.4
psycopg2-binary
python-jose[cryptography]       
passlib[bcrypt]                 
python-dotenv                  
alembic             
mysql-connector-python            
